Project 03_04 Finally Finished
After dealing with myself for most of this project in terms of how I EXACTLY prefer the appearance to look, I was able to satisfy myself with pretty much how the entire webpage looks. I even added comments as well which I tried to do for every single style rule as these were added directly by myself. 

I ended up taking a lot longer to finish this than I was expecting, but I'm decently proud of the result and I hope you are as well.